// src/types/hooks.ts
export interface Film {
    title: string;
    episode_id: number;
    opening_crawl: string;
    director: string;
    producer: string;
    release_date: string;
    characters: string[];
    planets: string[];
    starships: string[];
    vehicles: string[];
    species: string[];
    created: string;
    edited: string;
    url: string;
}

export type SortField = 'title' | 'episode_id' | 'producer' | 'characters';
export type SortDirection = 'asc' | 'desc';